806: Add multisig controller and comprehensive documentation
- Create multisigController with clean separation of concerns
- Refactor routes to use controller pattern
- Add detailed README for MultiSigService with usage examples
- Document proposal lifecycle and security considerations
- Include troubleshooting guide and error codes reference
- Add architecture overview connecting backend and on-chain contract
